fix(monitoring): Replace float by int on PUT /challenges

+++ Application/pac-back/src/monitoring/app/database/interactors/Monitoring/challenges.py
@@ -65,7 +65,7 @@ class MonitoringChallengesInteractor:
 
     @staticmethod
     async def update_challenge_status(
-        challenge_id: UUID, status: ChallengeStatus, score: float
+        challenge_id: UUID, status: ChallengeStatus, score: int
     ) -> Challenge | None:
         supabase: Client = get_db()
         response = (
